admin-service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.1
flask-helmet==1.0.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-core==0.3.65
langchain-ollama==0.3.3
langchain-openai==0.3.22
langchain-pinecone==0.2.8
langchain-tests==0.3.20
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
narwhals==1.42.0
nltk==3.9.1
numpy==2.3.0
ollama==0.5.1
openai==1.86.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
pinecone==7.0.2
pinecone-client==6.0.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
pip==25.1
pluggy==1.6.0
propcache==0.3.2
protobuf==6.31.1
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pygraphviz==1.14
pyjnius==1.6.1
pytest==8.4.0
pytest-asyncio==0.26.0
pytest-benchmark==5.1.0
pytest-codspeed==3.2.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
streamlit==1.45.1
syrupy==4.9.1
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
vcrpy==7.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
